DD-70 refactor : praise 도메인에서 sendpraise, receivepraise로 분리:
