+++ src/main/java/com/dodok/honeypot/domain/badge/helper/CheckSendPraiseBadgeHelper.java
@@ -6,7 +6,7 @@ import com.dodok.honeypot.domain.badge.repository.BadgeCompleteRepository;
 import com.dodok.honeypot.domain.badge.repository.BadgeRepository;
 import com.dodok.honeypot.domain.member.error.MemberErrorCode;
 import com.dodok.honeypot.domain.member.repository.MemberRepository;
-import com.dodok.honeypot.domain.praise.repository.SendPraiseRepository;
+import com.dodok.honeypot.domain.sendpraise.repository.SendPraiseRepository;
 import com.dodok.honeypot.global.error.exception.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
